Some cleanup and extract last view functions to files
