+++ CrusherChess/src/hash.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "zobristkeys.h"
+#include "zobristKeys.h"
 
 
 /****************************************\
