+++ CrusherChess/src/perft.h
@@ -1,0 +1,3 @@
+#pragma once
+void perft_driver(int depth);
+void perft_test(int depth);
